% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binb.R
\name{metropolis}
\alias{metropolis}
\alias{iqss}
\title{Binb is not Beamer - PDF Presentation Themes}
\usage{
metropolis(toc = FALSE, slide_level = 2, incremental = FALSE,
  fig_width = 10, fig_height = 7, fig_crop = TRUE,
  fig_caption = TRUE, dev = "pdf", df_print = "default",
  fonttheme = "default", highlight = "tango", keep_tex = FALSE,
  latex_engine = "xelatex", citation_package = c("none", "natbib",
  "biblatex"), includes = NULL, md_extensions = NULL,
  pandoc_args = NULL)

iqss(toc = FALSE, slide_level = 3, incremental = FALSE,
  fig_width = 10, fig_height = 7, fig_crop = TRUE,
  fig_caption = TRUE, dev = "pdf", df_print = "default",
  fonttheme = "default", highlight = "haddock", keep_tex = FALSE,
  latex_engine = "xelatex", citation_package = c("none", "natbib",
  "biblatex"), includes = NULL, md_extensions = NULL,
  pandoc_args = NULL)
}
\arguments{
\item{toc}{A logical variable defaulting to \code{FALSE}.}

\item{slide_level}{A numeric variable defaulting to two for
\sQuote{Metrpolis} and three for \sQuote{IQSS}.}

\item{incremental}{A logical variable defaulting to \code{FALSE}.}

\item{fig_width}{A numeric variable defaulting to ten.}

\item{fig_height}{A numeric variable defaulting to seven.}

\item{fig_crop}{A logical variable defaulting to \code{TRUE}.}

\item{fig_caption}{A logical variable defaulting to \code{TRUE}.}

\item{dev}{A character variable defaulting to \dQuote{pdf}.}

\item{df_print}{A character variable defaulting to \dQuote{default}.}

\item{fonttheme}{A character variable defaulting to \dQuote{default}.}

\item{highlight}{A character variable defaulting to \dQuote{tango}.}

\item{keep_tex}{A logical variable defaulting to \code{FALSE}.}

\item{latex_engine}{A character variable defaulting to \dQuote{xelatex}.}

\item{citation_package}{An optional character variable with possible value
\dQuote{none}, \dQuote{natbib} (the default), or \dQuote{biblatex}.}

\item{includes}{An optional character variable defaulting to \code{NULL}.}

\item{md_extensions}{An optional character variable defaulting to \code{NULL}.}

\item{pandoc_args}{An optional character variable defaulting to \code{NULL}.}
}
\value{
RMarkdown content processed for rendering.
}
\description{
Templates for RMarkdown-based Beamer PDF presentations supporting
the \sQuote{Metropolis} style by Matthias Vogelgesang and others,
and the \sQuote{IQSS} style by Ista Zahn and Gary King.
}
\details{
Note that not all options and customizations available at the LaTeX
level are implemented or supported yet.
}
\examples{
\dontrun{
library(rmarkdown)
draft("myslides.Rmd", template="metropolis", package="binb", edit=FALSE)
setwd("myslides")       ## template creates a new subdir
render("myslides.Rmd")
}
}
\author{
Dirk Eddelbuettel
}
